+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - added some error message for missing injected service
 - moved CHANGE_ROUTE Event-Type from Coco.Evnet to Coco.RouterEvent
 - moved FIRE_ROUTE Event-Type from Coco.Evnet to Coco.RouterEvent
+- moved SHOW_VIEW Event-Type from Coco.Evnet to Coco.RouterEvent
+- moved HIDE_VIEW Event-Type from Coco.Evnet to Coco.RouterEvent
 - added oldRoute object to Coco.RouterEvent
 
 ## 0.1.4